+++ packages/medical-domain/src/domains/repositories/medical/medical-data.service.ts
@@ -91,7 +91,6 @@ class MedicalDataService {
     pumpSettings: [],
     reservoirChanges: [],
     smbg: [],
-    uploads: [],
     warmUps: [],
     wizards: [],
     zenModes: [],
@@ -341,9 +340,6 @@ class MedicalDataService {
             break
           case DatumType.Smbg:
             this.medicalData.smbg.push(datum)
-            break
-          case DatumType.Upload:
-            this.medicalData.uploads.push(datum)
             break
           case DatumType.Wizard:
             this.medicalData.wizards.push(datum)