+++ packages/yourloops/test/integration/assert/patient-nav-bar.assert.ts
@@ -32,33 +32,36 @@ import moment from 'moment-timezone'
 import userEvent from '@testing-library/user-event'
 import { type Data, mockDataAPI } from '../mock/data.api.mock'
 
-const checkPatientNavBar = (patientNavBar: BoundFunctions<typeof queries>) => {
-  const dashboardTab = patientNavBar.getByText('Dashboard')
-  const dailyTab = patientNavBar.getByText('Daily')
-  const trendsTab = patientNavBar.getByText('Trends')
-  expect(dashboardTab).toBeVisible()
-  expect(dailyTab).toBeVisible()
-  expect(trendsTab).toBeVisible()
+const checkPatientNavBarCommon = (patientNavBar: BoundFunctions<typeof queries>) => {
+  expect(patientNavBar.getByRole('tablist')).toHaveTextContent('DashboardDailyTrendsDevice')
   expect(patientNavBar.getByText('Download report')).toBeVisible()
 }
 
-export const checkPatientNavBarAsHCP = () => {
+export const checkPatientNavBarAsHcp = () => {
   const patientNavBar = within(screen.getByTestId('patient-nav-bar'))
   expect(patientNavBar.getByTestId('subnav-patient-list')).toBeVisible()
   expect(patientNavBar.getByTestId('patient-dropdown')).toBeVisible()
-  checkPatientNavBar(patientNavBar)
+  expect(patientNavBar.getByRole('tablist')).toHaveTextContent('DashboardDailyTrendsTarget & alertsDevice')
+  expect(patientNavBar.getByText('Download report')).toBeVisible()
+}
+
+export const checkPatientNavBarAsHcpInPrivateTeam = () => {
+  const patientNavBar = within(screen.getByTestId('patient-nav-bar'))
+  expect(patientNavBar.getByTestId('subnav-patient-list')).toBeVisible()
+  expect(patientNavBar.getByTestId('patient-dropdown')).toBeVisible()
+  checkPatientNavBarCommon(patientNavBar)
 }
 
 export const checkPatientNavBarAsCaregiver = () => {
   const patientNavBar = within(screen.getByTestId('patient-nav-bar'))
   expect(patientNavBar.getByTestId('subnav-patient-list')).toBeVisible()
   expect(patientNavBar.getByTestId('patient-dropdown')).toBeVisible()
-  checkPatientNavBar(patientNavBar)
+  checkPatientNavBarCommon(patientNavBar)
 }
 
 export const checkPatientNavBarAsPatient = () => {
   const patientNavBar = within(screen.getByTestId('patient-nav-bar'))
-  checkPatientNavBar(patientNavBar)
+  checkPatientNavBarCommon(patientNavBar)
 }
 
 export const checkPatientDropdown = async (initialPatient: Patient, patientToSwitchTo: Patient) => {
@@ -73,7 +76,9 @@ export const checkPatientDropdown = async (initialPatient: Patient, patientToSwi
   const secondPatientName = `${patientToSwitchTo.profile.firstName} ${patientToSwitchTo.profile.lastName}`
   const secondaryHeaderRefreshed = await screen.findByTestId('patient-nav-bar')
   const secondPatientHeaderContent = `Patient${secondPatientName}Date of birth:${secondPatientDateOfBirth}Diabetes type:Type 1Gender:FemaleHbA1c:8.9% (11/21/2023)Email:patient2@diabeloop.frDashboardDailyTrendsTarget & alertsDeviceDownload report`
-  await waitFor(() => { expect(secondaryHeaderRefreshed).toHaveTextContent(secondPatientHeaderContent) })
+  await waitFor(() => {
+    expect(secondaryHeaderRefreshed).toHaveTextContent(secondPatientHeaderContent)
+  })
 
   fireEvent.mouseDown(within(await screen.findByTestId('patient-nav-bar')).getByText(patientToSwitchTo.profile.fullName))
   await userEvent.click(within(screen.getByRole('listbox')).getByText(initialPatient.profile.fullName))