+++ packages/yourloops/test/integration/use-cases/patient-data-visualisation.ts
@@ -92,6 +92,7 @@ export const testTrensdsDataVisualisationForHCP = async () => {
   await checkPatientStatisticsTrendsView()
   await checkTrendsStatsWidgetsTooltips()
 }
+
 export const testPatientNavBarForHcp = async () => {
   await checkPatientDropdown(patient1Info, patient2Info)
 }